move the sembast ride dao
